refactors tab menu builer to decouple from component
