+++ src/components/header/Header.jsx
@@ -1,26 +1,22 @@
 import React, { useState, useEffect } from 'react'
 import {
-  CardMedia,
   Box,
-  Grid,
   useMediaQuery,
-  SwipeableDrawer,
-  IconButton,
 } from '@mui/material'
-import MenuIcon from '@mui/icons-material/Menu'
 import useStyles from '../../theme/views/navbar.theme'
-import Adaptive_logo from '../../assets/images/Adaptive_logo_WHITE_PINK.png'
-import navbarHelper from '../../modules/navbarHelper'
-import DesktopNavbar from './navbar/DesktopNavbar'
-import MobileNavbar from './navbar/MobileNavbar'
+import buildNavigationMenu from '../../modules/buildNavigationMenu'
+import DesktopNavbar from './navigation/DesktopNavbar'
+import MobileNavbar from './navigation/MobileNavbar'
+import NavTab from './navigation/NavTab'
+import tabsData from '../../assets/data/tabsData'
 
 const Navbar = () => {
   const classes = useStyles()
   const isMobile = useMediaQuery('(max-width:1700px)')
-  const [tabs, setTabs] = useState([])
-
+  const [tabs, setTabs] = useState([])  
+  
   useEffect(() => {
-    setTabs(navbarHelper.buildTabs())
+    setTabs(buildNavigationMenu(tabsData, NavTab))
   }, [isMobile])
 
   return (
