+++ src/components/header/navigation/MobileNavbar.jsx
@@ -1,9 +1,6 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState } from 'react'
 import {
-  CardMedia,
   Box,
-  Grid,
-  useMediaQuery,
   SwipeableDrawer,
   IconButton,
 } from '@mui/material'
